+++ utils.hpp
@@ -10,9 +10,10 @@ void exportVariable(std::string variable, std::string value);
 
 std::chrono::nanoseconds GetDuation(std::function<void()> func);
 
-double sequentialPI(const int numSamples);
+double sequentialPI(const unsigned long long numSamples);
 
-double parallelPI(const int numTotalSamples, const int numChunks = 8);
+double parallelPI(const unsigned long long numTotalSamples,
+                  const unsigned long long numChunks = 8);
 
 char* sequentialDNA(int n, char* dna);
